Introduce BlockReader helper class.
Refactor the reading part to split out some logic into a BlockReader
class.  In particular, this class has now the logic to query for the
txid(s) associated to a given block height.

With this, the performRead function is simpler, and also this logic
can be reused in the future.